Combine ChromeOS screen_lock with existing Osquery MacOS screenlock table, specify idle time for system_state
