+++ ee/fleetd-chrome/src/db.ts
@@ -9,7 +9,7 @@ import TableNetworkInterfaces from "./tables/network_interfaces";
 import TableOsqueryInfo from "./tables/osquery_info";
 import TableOSVersion from "./tables/os_version";
 import TablePrivacyPreferences from "./tables/privacy_preferences";
-import TableScreenLock from "./tables/screen_lock";
+import TableScreenLock from "./tables/screenlock";
 import TableSystemInfo from "./tables/system_info";
 import TableSystemState from "./tables/system_state";
 import TableUsers from "./tables/users";
