+++ ee/fleetd-chrome/src/tables/system_state.ts
@@ -9,8 +9,11 @@ export default class TableSystemState extends Table {
       chrome.idle.getAutoLockDelay(resolve)
     )) as number;
 
+    // Idle time is set to 20% of the user's autolock time or defaults to 30 seconds
+    const idleStateDelay = autoLockDelay > 0 ? 0.2 * autoLockDelay : 30;
+
     const idleState = (await new Promise((resolve) =>
-      chrome.idle.queryState(autoLockDelay, resolve)
+      chrome.idle.queryState(idleStateDelay, resolve)
     )) as string;
 
     return [{ idle_state: idleState }];
