--- schema/tables/screen_lock.yml
@@ -1,18 +1,0 @@
-name: screen_lock
-platforms:
-  - chrome
-description: Gets the time, in seconds, it takes until the screen is locked automatically while idle. Returns a zero duration if the screen is never locked automatically.
-examples: >-
-  Gets the time, in seconds, it takes until the screen is locked automatically while idle. 
-
-  ```
-
-  SELECT delay from screen_lock;
-
-  ```
-columns:
-  - name: delay
-    type: integer
-    description: Time, in seconds, it takes until the screen is locked automatically while idle.
-    required: false
-evented: false
